+++ include/coresql/timestamp.hpp
@@ -2,7 +2,7 @@
 #include "core.hpp"
 
 namespace coresql::timestamps {
-Type type();
+const Type& type();
 Value value(std::int64_t unix_microseconds);
 std::int64_t microseconds(const Value&);
 void install(Registry&);